Shorten certs cli (#2611)
