+++ agent/provider/src/cli/keystore.rs
@@ -2,13 +2,14 @@ use crate::cli::println_conditional;
 use crate::rules::{CertWithRules, RulesManager};
 use crate::startup_config::ProviderConfig;
 use chrono::{DateTime, SecondsFormat, Utc};
-use itertools::Itertools;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 use std::path::PathBuf;
+use std::vec;
 use structopt::StructOpt;
 use ya_manifest_utils::keystore::{
     AddParams, AddResponse, Cert, Keystore, RemoveParams, RemoveResponse,
 };
+use ya_manifest_utils::short_cert_ids::shorten_cert_ids;
 use ya_utils_cli::{CommandOutput, ResponseTable};
 
 /// Manage trusted keys
@@ -201,55 +202,14 @@ impl CertTableBuilder {
     }
 
     pub fn build(self) -> CertTable {
-        const DIGEST_PREFIX_LENGTHS: [usize; 3] = [8, 32, 128];
+        let long_ids: Vec<String> = self.entries.iter().map(|e| e.cert.id()).collect();
 
-        // hard-code support for the use of the entire signature, regardless of its size,
-        // ensure all prefixes are no longer than the signature, and remove duplicates.
-        //
-        // these are, by construction, sorted smallest to largest.
-        let prefix_lengths = |id_len| {
-            DIGEST_PREFIX_LENGTHS
-                .iter()
-                .map(move |&n| std::cmp::min(n, id_len))
-                .chain(std::iter::once(id_len))
-                .dedup()
-        };
-
-        let mut prefix_uses = HashMap::<String, u32>::new();
-        for cert in &self.entries {
-            for len in prefix_lengths(cert.cert.id().len()) {
-                let mut prefix = cert.cert.id();
-                prefix.truncate(len);
-
-                *prefix_uses.entry(prefix).or_default() += 1;
-            }
-        }
-
-        let mut ids = Vec::new();
-        for cert in &self.entries {
-            for len in prefix_lengths(cert.cert.id().len()) {
-                let mut prefix = cert.cert.id();
-                prefix.truncate(len);
-
-                let usages = *prefix_uses
-                    .get(&prefix)
-                    .expect("Internal error, unexpected prefix");
-
-                // the longest prefix (i.e. the entire fingerprint) will be unique, so
-                // this condition is guaranteed to execute during the last iteration,
-                // at the latest.
-                if usages == 1 {
-                    ids.push(prefix);
-                    break;
-                }
-            }
-        }
-
-        let mut values = Vec::new();
-        for (id_prefix, cert) in ids.into_iter().zip(self.entries.into_iter()) {
-            let not_after_formatted = date_to_str(&cert.cert.not_after());
+        let short_ids = shorten_cert_ids(&long_ids);
+        let mut values = vec![];
+        for (entry, short_id) in self.entries.into_iter().zip(short_ids) {
+            let not_after_formatted = date_to_str(&entry.cert.not_after());
             values
-                .push(serde_json::json! { [ id_prefix, cert.cert.type_name(), not_after_formatted, cert.cert.subject(), cert.format_outbound_rules()] });
+                .push(serde_json::json! { [ short_id, entry.cert.type_name(), not_after_formatted, entry.cert.subject(), entry.format_outbound_rules()] });
         }
 
         let columns = vec![
